貸出登録のバリデーションテストを別ファイルに移行
